Use totally new approach, for much better performance
Instead of copying the entire original image from the DVD first, only to
replace almost all data with the decoded VOBs afterwards, only copy the
parts we actually keep unaltered.

This is achieved by compiling a list of all altered and unaltered
fragments; and then directly composing the final image from the decoded
files as well as the other fragments copied directly from DVD.

This is much much faster, as it not only saves the unnecessary copy of
most of the DVD; but also, compiling the image by appending the
fragments sequentially for some reason is much faster than replacing
fragments in the complete image.

This is still not optimal though, as the decoded files are still saved
to disk first, and only copied to the final image afterwards. This could
be avoided by getting a list of all VOBs first, and then extracting them
one by one directly to the target image. The latter is presently not
possible with dvdbackup, but could be done with vobcopy -- however,
vobcopy appears even less maintained, and the code is horrible -- so I'm
wary of prompting people to install or use it...