Check for pre-existing output file
Instead of appending to a possibly already existing .iso file --
resulting in a useless mess -- check for this situation early, and bail
out cleanly.

Note: it would be nice to also add an -f option, to remove an existing
file instead of bailing out. We need to implement option parsing for
that though...